ml/tgis/requirements.txt
numpy>=1.26
pandas>=2.2
scikit-learn>=1.4
pyyaml>=6.0
psycopg[binary]>=3.2
python-dotenv>=1.0
requests>=2.32
Pillow>=10.0
